Breaking change: fixing config directory. Also, 1.0.0 (#482)
* Breaking change: fixing config directory

Apparently release builds were failing to run on some systems. For
example, when running AppImage on linux, the previously used call the
get resource paths would point to a temp dir in `/tmp`. This would lead
to errors attempting to create the DB file.

The new approach points to `~/.config/iron-wallet`.

Since this is a breaking change, I also took this opportunity to rename
the identifier from `com.naps62.iron` to `iron-wallet`, since that's
where the directory name will come from.

I'm also happy with the idea of not making `1.0` be a very big deal, and
instead follow semver religiously. So there we go, we reached 1.0!

* code review

* code review

* code review

* code review