Refactor the way github checks are skipped (#1029)
The original approach meant workflows never actually ran when certain
files weren't changed.
this means we couldn't set those workflows as required, because the
pending check would never happen

the new approach is sort of a workaround for that. we use a dedicated
action to check if files were changed, instead of the top level workflow
triggers
this makes it so that the workflow technically runs, but ends up
skipping all the work if those files aren't changed