Add Route guard (#1434)
This PR adds a route guard to protect certain routes under specific
conditions.

Example: If we are on the Anvil network and viewing the `Explorer
Addresses` route, switching to another network should not keep us on
that page (since it would attempt to load addresses from the not
permitted network). Instead, we redirect the user to the `Account` page.